+++ src/main/java/com/stock/flex/resource/response/UserResponse.java
@@ -1,4 +1,4 @@
-package com.stock.flex.resource.request;
+package com.stock.flex.resource.response;
